refactor: Modal과 Overlay 폴더 구조 변경
