+++ src/components/common/Overlay/Overlay.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Overlay.module.css';
+import styles from './css/Overlay.module.css';
 
 export default function Overlay({ onClose }) {
   return <div onClick={onClose} className={styles.overlay}></div>;
